[rename] dto 변경 (#66)
