+++ src/main/java/hyangyu/server/dto/event/DisplayResponseDto.java
@@ -1,11 +1,11 @@
-package hyangyu.server.dto.myPage;
+package hyangyu.server.dto.event;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
 @Data
 @AllArgsConstructor
-public class MyDisplayResponseDto {
+public class DisplayResponseDto {
     int status;
-    MyDisplayDto data;
+    DisplayDto data;
 }
