+++ src/main/java/hyangyu/server/dto/event/FairResponseDto.java
@@ -1,11 +1,11 @@
-package hyangyu.server.dto.myPage;
+package hyangyu.server.dto.event;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
 @Data
 @AllArgsConstructor
-public class MyFairResponseDto {
+public class FairResponseDto {
     int status;
-    MyFairDto data;
+    FairDto data;
 }
