+++ src/main/java/hyangyu/server/dto/event/MyPageResponseDto.java
@@ -1,4 +1,4 @@
-package hyangyu.server.dto.myPage;
+package hyangyu.server.dto.event;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
